Lesson3: fixed the first two comments
